HTTP Requestが通りました記念

+++ app/src/main/java/com/syuheifujita/android_weather_app/model/Main.kt
@@ -4,12 +4,12 @@ import java.io.Serializable
 
 data class Main (
 
-    val temp: Int,
-    val pressure: Int,
+    val temp: Double,
+    val pressure: Double,
     val humidity: Int,
-    val temp_min: Int,
-    val temp_max: Int,
-    val sea_level: Int,
-    val gmd_level: Int
+    val temp_min: Double,
+    val temp_max: Double,
+    val sea_level: Double,
+    val gmd_level: Double
 
 ): Serializable
